Getting test environment setup
