+++ .eslintrc.js
@@ -6,5 +6,8 @@
  */
 
 module.exports = {
-  extends: './node_modules/eslint-config-airbnb-base/index.js'
+  extends: './node_modules/eslint-config-airbnb-base/index.js',
+  env: {
+    "jest": true
+  }
 };
