+++ client/src/prodDesc.jsx
@@ -5,7 +5,7 @@ class ProdDesc extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      currentProduct: this.getProductByName('Bolo Tie')
+      currentProduct: this.getProductById(22)
     }
   }
